+++ tests/headerTableInput.js
@@ -5,7 +5,7 @@ import h from "./hyperscript";
 const value = (
   <value>
     <document>
-      <table headless={true}>
+      <table headless={false}>
         <tr>
           <td>
             <p>Col 0, Row 0</p>
@@ -16,7 +16,7 @@ const value = (
         </tr>
         <tr>
           <td>
-            <p key="_cursor_">Col 0, Row 1</p>
+            <p>Col 0, Row 1</p>
           </td>
           <td>
             <p>Col 1, Row 1</p>
@@ -31,7 +31,6 @@ const value = (
           </td>
         </tr>
       </table>
-      <p key="_cursor_2_">An extraneous paragraph</p>
     </document>
   </value>
 );